+++ apps/webapp/charmClient/apis/subscriptionApi.ts
@@ -12,9 +12,11 @@ import type { CouponDetails } from '@packages/lib/subscription/getCouponDetails'
 import type { CreateProSubscriptionRequest, SubscriptionPaymentIntent } from '@packages/lib/subscription/interfaces';
 import type { UpdatePaymentMethodRequest } from '@packages/lib/subscription/updatePaymentMethod';
 import type { UpdateSubscriptionRequest } from '@packages/lib/subscription/updateProSubscription';
-import type { UpgradeSubscriptionRequest } from '@packages/lib/subscription/upgradeProSubscription';
-import type { CreateSubscriptionContributionRequest } from '@packages/spaces/createSpaceContribution';
-import type { SubscriptionReceipt } from '@packages/spaces/getSubscriptionReceipts';
+import type { UpgradeSubscriptionRequest as UpgradeSubscriptionRequestV2 } from '@packages/lib/subscription/upgradeProSubscription';
+import type { CreateSubscriptionContributionRequest } from '@packages/subscriptions/createSubscriptionContribution';
+import type { DowngradeSubscriptionRequest } from '@packages/subscriptions/downgradeSubscription';
+import type { SubscriptionReceipt } from '@packages/subscriptions/getSubscriptionReceipts';
+import type { UpgradeSubscriptionRequest } from '@packages/subscriptions/upgradeSubscription';
 
 export class SubscriptionApi {
   createSubscription(spaceId: string, payload: CreateProSubscriptionRequest) {
@@ -41,7 +43,7 @@ export class SubscriptionApi {
     return http.POST<CouponDetails | null>(`/api/spaces/${spaceId}/validate-discount`, payload);
   }
 
-  upgradeSpaceSubscription(spaceId: string, payload: UpgradeSubscriptionRequest) {
+  upgradeSpaceSubscription(spaceId: string, payload: UpgradeSubscriptionRequestV2) {
     return http.PUT<void>(`/api/spaces/${spaceId}/upgrade-subscription`, payload);
   }
 
@@ -61,7 +63,15 @@ export class SubscriptionApi {
     return http.POST(`/api/spaces/${spaceId}/subscriptions/contribution`, payload);
   }
 
-  getSubscriptionContributions(spaceId: string) {
-    return http.GET<SubscriptionReceipt[]>(`/api/spaces/${spaceId}/subscriptions/contribution`);
+  getSubscriptionReceipts(spaceId: string) {
+    return http.GET<SubscriptionReceipt[]>(`/api/spaces/${spaceId}/subscriptions/receipts`);
+  }
+
+  upgradeSubscription(spaceId: string, payload: UpgradeSubscriptionRequest) {
+    return http.POST<void>(`/api/spaces/${spaceId}/subscriptions/upgrade`, payload);
+  }
+
+  downgradeSubscription(spaceId: string, payload: DowngradeSubscriptionRequest) {
+    return http.POST<void>(`/api/spaces/${spaceId}/subscriptions/downgrade`, payload);
   }
 }
