--- apps/webapp/pages/api/spaces/[id]/upgrade-tier.ts
@@ -1,26 +1,0 @@
-import { onError, onNoMatch, requireSpaceMembership, requireUser } from '@packages/lib/middleware';
-import { withSessionRoute } from '@packages/lib/session/withSession';
-import type { UpgradeSubscriptionTierRequest } from '@packages/lib/subscription/upgradeSubscriptionTier';
-import { upgradeSubscriptionTier } from '@packages/lib/subscription/upgradeSubscriptionTier';
-import type { NextApiRequest, NextApiResponse } from 'next';
-import nc from 'next-connect';
-
-const handler = nc<NextApiRequest, NextApiResponse>({ onError, onNoMatch });
-
-handler
-  .use(requireUser)
-  .use(requireSpaceMembership({ adminOnly: true, spaceIdKey: 'id' }))
-  .post(upgradeTierController);
-
-async function upgradeTierController(req: NextApiRequest, res: NextApiResponse<string>) {
-  const { id: spaceId } = req.query as { id: string };
-
-  await upgradeSubscriptionTier({
-    ...(req.body as UpgradeSubscriptionTierRequest),
-    spaceId
-  });
-
-  res.status(200).end();
-}
-
-export default withSessionRoute(handler);
