+++ packages/lib/tsconfig.json
@@ -6,9 +6,16 @@
     // composite is useful for referenced packages. see https://www.typescriptlang.org/docs/handbook/project-references.html#composite
     "composite": true
   },
-  "include": ["src/**/*.ts","src/**/*.tsx"],
+  "include": [
+    "src/**/*.ts",
+    "src/**/*.tsx"
+  ],
   "paths": {
-    "lib/*": ["./src/*"],
-    "@packages/lib/*": ["./src/*"],
+    "lib/*": [
+      "./src/*"
+    ],
+    "@packages/lib/*": [
+      "./src/*"
+    ],
   }
 }
