+++ packages/subscriptions/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "@packages/subscriptions",
+  "version": "0.0.0",
+  "description": "",
+  "type": "module",
+  "private": true,
+  "scripts": {
+    "test": ""
+  },
+  "exports": {
+    "./*": "./src/*.ts"
+  },
+  "dependencies": {
+    "@packages/utils": "^1.0.0",
+    "viem": "^2.29.2"
+  }
+}
