+++ packages/subscriptions/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../../tsconfig.base.json",
+  "compilerOptions": {
+    "outDir": "dist",
+    "rootDir": "src",
+    // composite is useful for referenced packages. see https://www.typescriptlang.org/docs/handbook/project-references.html#composite
+    "composite": true
+  }
+}
